Fixed. 디렉토리 구조 수정 및 파일 위치 조정
