+++ src/main/java/io/github/sunday/devfolio/entity/table/profile/Follow.java
@@ -1,4 +1,4 @@
-package io.github.sunday.devfolio.model;
+package io.github.sunday.devfolio.entity.table.profile;
 
 import lombok.*;
 import jakarta.persistence.*;
