+++ src/main/java/io/github/sunday/devfolio/entity/table/profile/Training.java
@@ -1,4 +1,4 @@
-package io.github.sunday.devfolio.model;
+package io.github.sunday.devfolio.entity.table.profile;
 
 import lombok.*;
 import jakarta.persistence.*;
@@ -6,30 +6,26 @@ import java.time.LocalDate;
 import java.util.Objects;
 
 /**
- * 스킬(사용 언어 및 툴)(Skill) 정보를 저장하는 엔티티.
+ * 교육/트레이닝(Training) 정보를 나타내는 엔티티입니다.
  */
 @Entity
-@Table(name = "skills")
+@Table(name = "training")
 @Getter @Setter @NoArgsConstructor @AllArgsConstructor @Builder
-public class Skill {
-    /** 스킬 고유 식별자 (PK) */
+public class Training {
+    /** 교육 고유 식별자 (PK) */
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "skill_idx")
-    private Long skillIdx;
+    @Column(name = "training_idx")
+    private Long trainingIdx;
 
-    /** 스킬이 속한 이력서 참조 */
+    /** 교육이 속한 이력서 참조 */
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "resume_idx", nullable = false)
     private Resume resume;
 
-    /** 스킬 이름 */
-    @Column(name = "name", length = 200)
-    private String name;
-
-    /** 숙련도 레벨 (예: 1~5) */
-    @Column(name = "level")
-    private Integer level;
+    /** 교육/트레이닝 제목 */
+    @Column(name = "training_name", length = 200)
+    private String trainingName;
 
     /** 시작 일자 */
     @Column(name = "start_date")
@@ -39,17 +35,21 @@ public class Skill {
     @Column(name = "end_date")
     private LocalDate endDate;
 
+    /** 상세 내용 */
+    @Column(name = "description", columnDefinition = "TEXT")
+    private String description;
+
     /** 객체 동등성 비교 */
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
-        Skill that = (Skill) o;
-        return Objects.equals(skillIdx, that.skillIdx);
+        Training that = (Training) o;
+        return Objects.equals(trainingIdx, that.trainingIdx);
     }
     /** 해시코드계산 */
     @Override
     public int hashCode() {
-        return Objects.hash(skillIdx);
+        return Objects.hash(trainingIdx);
     }
 }
