+++ src/main/java/io/github/sunday/devfolio/repository/FollowRepository.java
@@ -1,6 +1,6 @@
 package io.github.sunday.devfolio.repository;
 
-import io.github.sunday.devfolio.model.Follow;
+import io.github.sunday.devfolio.entity.table.profile.Follow;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
